Supabase対応

+++ src/app/admin/categories/_components/CategoryForm.tsx
@@ -1,22 +1,92 @@
 "use client"
-import { useEffect, useState } from "react"
+import { useEffect } from "react"
 import { useRouter } from "next/navigation"
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession"
+import { useForm, SubmitHandler } from "react-hook-form"
+// import useSWR from "swr"
+import { Category } from "@/app/_types/Category";
+import { useFetch } from '@/app/_hooks/useFetch'
+
+type CategoryResponse = {
+  category: Category
+}
+
+type Inputs = {
+  name: string
+}
+
+// const fetcher = async ([url, token]: [string, string]) => {
+//   const res = await fetch(url, {
+//     headers: {
+//       'Content-Type': 'application/json',
+//       Authorization: token,
+//     },
+//   })
+//   if (!res.ok) {
+//     throw new Error('データの取得に失敗しました')
+//   }
+//   return res.json()
+// }
 
 export default function CategoryForm( {id}: {id: string | undefined}) {
-  const [name, setName] = useState('')
+  // const [name, setName] = useState('')
+  const { register, handleSubmit, formState: { errors, isSubmitting }, reset } = useForm<Inputs>({
+    defaultValues: {
+      name: '',
+    }
+  });
   const router = useRouter()
+  const { token } = useSupabaseSession()
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
+  // useEffect(() => {
+  //   if (!token) return
+  //   if (!id) return
+  //   const fetcher = async () => {
+  //     try{
+  //       const res = await fetch(`/api/admin/categories/${id}`,{
+  //         headers: {
+  //           'Content-Type': 'application/json',
+  //           Authorization: token,
+  //         },
+  //       })
+  //       const data = await res.json()
+  //         // setName(data.category.name)
+  //         reset({
+  //           name: data.category.name,
+  //         })
+  //     } catch (error) {
+  //       console.error('エラーが発生しました:', error)
+  //     }
+  //   }
+  //   fetcher()
+  // }, [id, token, reset])
+
+  const { data, error, isLoading } = useFetch<CategoryResponse>(
+    id ? `/api/admin/categories/${id}` : null
+  )
+
+  useEffect(() => {
+    if (data?.category) {
+      reset({
+        name: data.category.name,
+      })
+    }
+  }, [data, reset])
+
+  // const handleSubmit = async (e: React.FormEvent) => {
+  const onSubmit: SubmitHandler<Inputs> = async (data) => {
+    // e.preventDefault()
+    if (!token) return
     try{
       if (!id) {
         await fetch(`/api/admin/categories`, {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json',
+            Authorization: token,
           },
           body: JSON.stringify({
-            name,
+            name: data.name,
           }),
         })
         alert('登録しました')
@@ -25,9 +95,10 @@ export default function CategoryForm( {id}: {id: string | undefined}) {
           method: 'PUT',
           headers: {
             'Content-Type': 'application/json',
+            Authorization: token,
           },
           body: JSON.stringify({
-            name,
+            name: data.name,
           }),
         })
         alert('更新しました')
@@ -39,9 +110,14 @@ export default function CategoryForm( {id}: {id: string | undefined}) {
   }
   
   const handleDelete = async () => {
+    if (!token) return
     try { 
       await fetch(`/api/admin/categories/${id}`, {
         method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: token,
+        },
       })
       router.push(`/admin/categories`)
       alert('削除しました')
@@ -50,38 +126,32 @@ export default function CategoryForm( {id}: {id: string | undefined}) {
     }
   }
 
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value } = e.target
-    setName(value)
+  if (id && isLoading) {
+    return <div>読み込み中...</div>
   }
 
-  useEffect(() => {
-    if (!id) return
-    const fetcher = async () => {
-      try{
-        const res = await fetch(`/api/admin/categories/${id}`)
-        const data = await res.json()
-        setName(data.category.name)
-      } catch (error) {
-        console.error('エラーが発生しました:', error)
-      }
-    }
-    fetcher()
-  }, [id])
+  if (id && error) {
+    return <div>エラーが発生しました</div>
+  }
+
+  // const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  //   const { value } = e.target
+  //   setName(value)
+  // }
 
   return (
-    <div>
+    <form onSubmit={handleSubmit(onSubmit)}>
       {id && <h1 className="text-xl font-bold pb-4">カテゴリー編集</h1>}
       {!id && <h1 className="text-xl font-bold pb-4">カテゴリー新規作成</h1>}
       <div className="mb-4">
         <div className="">カテゴリー名</div>
-        <input id="name" value={name} name="name" type="text" className="w-full border border-gray-300 rounded h-8 p-6" onChange={handleChange}/>
+        <input id="name" type="text" className="w-full border border-gray-300 rounded h-8 p-6" {...register('name', {required: 'カテゴリー名を入力してください'})}/>
+        {errors.name && <span className="text-red-500">{errors.name.message}</span>}
       </div>
       <div className="flex justify-left">
-        <button type='submit' className="bg-gray-800 text-white font-bold rounded-lg px-6 py-2 mr-4 hover:cursor-pointer" onClick={handleSubmit}>{id ? '更新' : '新規作成'}</button>
+        <button type='submit' className="bg-gray-800 text-white font-bold rounded-lg px-6 py-2 mr-4 hover:cursor-pointer" disabled={isSubmitting}>{isSubmitting ? '更新中...' : id ? '更新' : '新規作成'}</button>
         {id && <button type='button' className="bg-gray-200 text-gray-800 font-bold rounded-lg px-6 py-2 hover:cursor-pointer" onClick={handleDelete}>削除</button>}
       </div>
-    </div>
+    </form>
   )
 }
